+++ cocos/scripting/lua/bindings/CCLuaStack.cpp
@@ -191,7 +191,7 @@ void LuaStack::addLuaLoader(lua_CFunction func)
     
     // insert loader into index 2
     lua_pushcfunction(_state, func);                                   /* L: package, loaders, func */
-    for (int i = lua_objlen(_state, -2) + 1; i > 2; --i)
+    for (int i = static_cast<int>(lua_objlen(_state, -2) + 1); i > 2; --i)
     {
         lua_rawgeti(_state, -2, i - 1);                                /* L: package, loaders, func, function */
         // we call lua_rawgeti, so the loader table now is at -3
@@ -411,7 +411,7 @@ int LuaStack::executeFunction(int numArgs)
     int ret = 0;
     if (lua_isnumber(_state, -1))
     {
-        ret = lua_tointeger(_state, -1);
+        ret = static_cast<int>(lua_tointeger(_state, -1));
     }
     else if (lua_isboolean(_state, -1))
     {
